# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Yummy API",
    "description": "API documentation for team pet project \"Yummy\"\n",
    "contact": {
      "name": "Oleksandr Kochenko",
      "url": "https://oleksandrkochenko.github.io/CV/",
      "email": "kochenko.fop@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://yummy-pet-project.onrender.com"
    },
    {
      "url": "http://localhost:4000/"
    },
    {
      "url": "https://virtserver.swaggerhub.com/KOCHENKOFOP/YummyApi/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
...
  "paths": {
...
    "/api/categories": {
      "get": {
        "tags": [
          "API / Categories"
        ],
        "summary": "Fetch categories",
        "operationId": "getCategories",
        "responses": {
          "200": {
            "description": "Categories array",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Not valid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/front-page": {
      "get": {
        "tags": [
          "API / Recipes"
        ],
        "summary": "Fetch front page recipes",
        "operationId": "getFrontPage",
        "responses": {
          "200": {
            "description": "Front page recipes data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FrontPage"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Not valid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes": {
      "get": {
        "tags": [
          "API / Recipes"
        ],
        "summary": "Page of recipes by at least one of params - category, title or ingredient",
        "operationId": "getRecipesByParams",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "title value that need to be considered for filter",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "enum": [
                "cheesecake",
                "chicken",
                "soup",
                "tart"
              ]
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "category value that need to be considered for filter",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "enum": [
                "Beef",
                "Chicken",
                "Dessert",
                "Vegetarian"
              ]
            }
          },
          {
            "name": "ingredient",
            "in": "query",
            "description": "ingredient id that need to be considered for filter",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "enum": [
                "640c2dd963a319ea671e365b",
                "640c2dd963a319ea671e3666",
                "640c2dd963a319ea671e368a"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of recipes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipesPage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Not valid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "API / Recipes"
        ],
        "summary": "Add new recipe",
        "operationId": "addNewRecipe",
        "requestBody": {
          "description": "New recipe object",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AddNewRecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Page of recipes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeItem"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Not valid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/{id}": {
      "get": {
        "tags": [
          "API / Recipes"
        ],
        "summary": "Fetch recipe by id",
        "description": "Returns a single recipe",
        "operationId": "getRecipeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of recipe to return",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "enum": [
                "6462a6cd4c3d0ddd28897f8e",
                "6462a8f74c3d0ddd28897fba",
                "6462a8f74c3d0ddd28897fb8"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Single recipe data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeItem"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Not valid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/my-recipes": {
      "get": {
        "tags": [
          "API / Recipes"
        ],
        "summary": "Fetch user's own recipes",
        "description": "Returns recipes added by user",
        "operationId": "getMyRecipes",
        "responses": {
          "200": {
            "description": "User's own recipes page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipesPage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Not valid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ingredients": {
      "get": {
        "tags": [
          "API / Ingredients"
        ],
        "summary": "Fetch ingredients",
        "operationId": "getIngredients",
        "responses": {
          "200": {
            "description": "Ingredients array",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Not valid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddNewRecipeRequest": {
        "required": [
          "category",
          "description",
          "ingredients",
          "instructions",
          "time",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of recipe",
            "example": "New Recipe"
          },
          "category": {
            "type": "string",
            "description": "Category of recipe",
            "example": "Seafood"
          },
          "description": {
            "type": "string",
            "description": "Short description of recipe",
            "example": "Very good description for a new recipe"
          },
          "instructions": {
            "type": "string",
            "description": "Instruction of recipe",
            "example": "This is a placeholder of an instruction for a new recipe"
          },
          "time": {
            "type": "string",
            "description": "Time of recipe preparation",
            "example": "30"
          },
          "ingredients": {
            "type": "array",
            "description": "array of ingredients id and measures",
            "items": {
              "$ref": "#/components/schemas/AddNewRecipeRequest_ingredients"
            }
          }
        }
      },
      "CategoryItem": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "category id",
            "example": "6462a6cd4c3d0ddd28897f8e"
          },
          "name": {
            "type": "string",
            "description": "category name",
            "example": "Beef"
          }
        }
      },
      "RecipeItem": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "recipe id",
            "example": "6462a6cd4c3d0ddd28897f8e"
          },
          "title": {
            "type": "string",
            "description": "recipe title",
            "example": "Beef Stroganoff"
          },
          "category": {
            "type": "string",
            "description": "recipe category",
            "example": "Beef"
          },
          "area": {
            "type": "string",
            "description": "recipe area",
            "example": "Britain"
          },
          "instructions": {
            "type": "string",
            "description": "recipe title",
            "example": "Take a beef, cut it and fry it on a pan."
          },
          "description": {
            "type": "string",
            "description": "recipe description",
            "example": "Beef Stroganoff is a famous recipe of fried beef"
          },
          "thumb": {
            "type": "string",
            "description": "recipe preview URL",
            "example": "https://www.themealdb.com/images/media/meals/2dsltq1560461468.jpg"
          },
          "preview": {
            "type": "string",
            "description": "recipe image URL",
            "example": "https://res.cloudinary.com/ddbvbv5sp/image/upload/v1678561360/s2xedheis7fera3ldvsb.jpg"
          },
          "time": {
            "type": "string",
            "description": "time of preparing",
            "example": "35"
          },
          "youtube": {
            "type": "string",
            "description": "recipe video URL",
            "example": "https://www.youtube.com/watch?v=C5n1fN8TGHs"
          },
          "tags": {
            "type": "array",
            "description": "recipe video URL",
            "items": {
              "type": "string",
              "description": "some tag",
              "example": "[\"Meat\",\"Beef\"]"
            }
          },
          "ingredients": {
            "type": "array",
            "description": "recipe ingredients",
            "items": {
              "$ref": "#/components/schemas/RecipeItem_ingredients"
            }
          }
        }
      },
      "IngredientItem": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ingredient id",
            "example": "640c2dd963a319ea671e3672"
          },
          "name": {
            "type": "string",
            "description": "ingredient name",
            "example": "Black Pepper"
          },
          "desc": {
            "type": "string",
            "description": "ingredient description",
            "example": "Black pepper (Piper nigrum) is a flowering vine in the family Piperaceae, cultivated for its fruit, known as a peppercorn, which is usually dried and used as a spice and seasoning."
          },
          "img": {
            "type": "string",
            "description": "ingredient image URL",
            "example": "https://res.cloudinary.com/ddbvbv5sp/image/upload/v1678564215/mj1ottbkds56ug62ptkz.png"
          }
        }
      },
      "FrontPage": {
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/CategoryItem"
          },
          "recipes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeItem"
            }
          }
        }
      },
      "RecipesPage": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 12
          },
          "recipes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeItem"
            }
          }
        }
      },
...
      "Unauthorized": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Unauthorized",
            "example": "Unauthorized"
          }
        }
      },
      "NotFound": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Not Found",
            "example": "Not Found"
          }
        }
      },
      "ServerError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Server Error",
            "example": "Server Error"
          }
        }
      },
...
      "AddNewRecipeRequest_ingredients": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "imgredient id",
            "example": "640c2dd963a319ea671e3669"
          },
          "measure": {
            "type": "string",
            "description": "imgredient measure",
            "example": "250 g"
          }
        }
      },
      "RecipeItem_ingredients": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ingredient id",
            "example": "640c2dd963a319ea671e3680"
          },
          "measure": {
            "type": "string",
            "description": "ingredient measure",
            "example": "2 tbsp"
          }
        },
        "description": "some ingredient"
      }
    },
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}